:art: onBtnRacingCountClicked() 모듈로 분리
